pyproject.toml
[project]
version = "3.3.0"
name = "Agents-eval"
description = "Assess the effectiveness of agentic AI systems across various use cases focusing on agnostic metrics that measure core agentic capabilities."
authors = [
    {name = "qte77", email = "qte@77.gh"}
]
readme = "README.md"
requires-python = "==3.13.*"
license = "bsd-3-clause"
dependencies = [
    # "agentops>=0.4.14",
    "httpx>=0.28.1",  # 2024
    "logfire>=4.24.0",
    "loguru>=0.7.3",  # 2024
    "markitdown[pdf]>=0.1.4",
    "pydantic>=2.12.5",
    # "pydantic-ai>=0.0.36",
    "pydantic-ai-slim[duckduckgo,openai,tavily]>=1.59.0",
    "pydantic-settings>=2.12.0",
    # Modern ROUGE, BLEU, text metrics (2024+ maintained)
    # DISABLED due to sentencepiece build issues
    # "torchmetrics[text]>=1.4.0",
    "scikit-learn>=1.8.0",  # F1, precision, recall, accuracy metrics
    "textdistance>=4.6.3",  # Multiple text similarity algorithms (2024)
    # # Fast C++ text similarity, replacement for unmaintained textdistance
    # "rapidfuzz>=3.14.3", 
    "networkx>=3.6.1",  # Graph analysis
    "scalene>=2.1.3",  # High-performance CPU, GPU, and memory profiler
    "arize-phoenix>=13.0.0",  # Local trace viewer via pip (replaces Docker-based Opik)
    "openinference-instrumentation-pydantic-ai>=0.1.12",  # PydanticAI auto-instrumentation, arizeai
]

[project.urls]
Documentation = "https://qte77.github.io/Agents-eval/"

[dependency-groups]
gui = [
    "streamlit>=1.54.0",
    "pyvis>=0.3.2",
]
wandb = [
    "weave>=0.52.28",
]
dev = [
    "complexipy>=5.2.0",
    "pymarkdownlnt>=0.9.35",
    "pyright>=1.1.408",
    "ruff>=0.15.1",
]
test = [
    "hypothesis>=6.151.4",
    "inline-snapshot>=0.31.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
    "reportlab>=4.4.10",  # for PDF generation
    "ruff>=0.9.2",
]
docs = [
    "griffe>=2.0.0",
    "mkdocs>=1.6.1",
    "mkdocs-awesome-pages-plugin>=2.9.3",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocs-section-index>=0.3.8",
    "mkdocstrings[python]>=1.0.3",
]

[tool.uv]
# package = true
default-groups = ["dev", "test", "gui", "docs"]
exclude-newer = "2026-02-15T00:00:00Z"


# MARK: Logging configurations


[tool.logfire]
ignore_no_config=true
send_to_logfire=false  # Was "if-token-present" - changed to disable pytest-logfire cloud API calls during test runs


# MARK: Testing and linting configurations



# MARK: Testing and linting configurations


[tool.pyright]
pythonVersion = "3.13"
include = ["src", "tests"]
typeCheckingMode = "strict"
extraPaths = ["./venv/lib/python3.13/site-packages"]
useLibraryCodeForTypes = true
reportMissingImports = true
reportMissingTypeStubs = false
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"

[tool.complexipy]
paths = ["src"]
max-complexity-allowed = 10
sort = "desc"

[tool.ruff]
target-version = "py313"
src = ["src", "tests"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
# ignore = ["E203"]  # Whitespace before ':'
unfixable = ["B"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E501"]  # Allow longer lines in test files

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
addopts = "--strict-markers -m 'not network and not benchmark'"
asyncio_mode = "auto"
# "function", "class", "module", "package", "session"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests",
    "benchmark: marks tests as benchmark tests",
    "network: marks tests requiring network access",
]

[tool.inline-snapshot]
format-command = "ruff format --stdin-filename {filename}"

[tool.coverage]

[tool.coverage.run]
include = [
    "src/**/*.py",
]
# omit = []
# branch = true

[tool.coverage.report]
show_missing = true
exclude_lines = [
    # 'pragma: no cover',
    'raise AssertionError',
    'raise NotImplementedError',
]
omit = [
    'env/*',
    'venv/*',
    '.venv/*',
    '*/virtualenv/*',
    '*/virtualenvs/*',
    '*/tests/*',
]

[tool.pymarkdown]
# Pymarkdown configuration for markdown linting
# Rules: https://github.com/jackdewinter/pymarkdown/blob/master/docs/rules.md
log.level = "CRITICAL"
# MD013: Line too long
plugins.md013.enabled = false

# MD025: Multiple top-level headings in the same document
plugins.md025.enabled = false

# MD032: Lists should be surrounded by blank lines
# Note: Autofix pending (scheduled after v1.0.0 release
# violations must be fixed manually before that
# https://pymarkdown.readthedocs.io/en/stable/plugins/rule_md032/
plugins.md032.enabled = true
# frontmatter extension suppresses violations as of v0.9.34
# the specific rules have toe be disabled instead
# extensions.front-matter.enabled = true
# MD041: First line should be a top-level heading
plugins.md041.enabled = false
# MD022: Headings need blank lines around
# plugins.md022.enabled = false
# MD003: Inconsistent heading styles (setext vs atx)
plugins.md003.enabled = false


# MARK: Version bumping configurations


[tool.bumpversion]
current_version = "3.3.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
commit = true
tag = true
allow_dirty = false
ignore_missing_version = false
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
message = "Bump version: {current_version} → {new_version}"
commit_args = ""

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "src/app/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "README.md"
search = "version-{current_version}-58f4c2"
replace = "version-{new_version}-58f4c2"

[[tool.bumpversion.files]]
filename = "CHANGELOG.md"
search = """
## [Unreleased]
"""
replace = """
## [Unreleased]

## [{new_version}] - {now:%Y-%m-%d}
"""
